# Conflicts:
#	rod/views.py

+++ rod/views.py
@@ -1,14 +1,13 @@
 from django.shortcuts import render, redirect
-from django.contrib.auth.models import User, auth
+from django.contrib.auth.models import  User, auth
 from .models import *
 from django.core.paginator import Paginator
 from .forms import DronUpdateForm
 from base64 import b64encode
 
-
 def home(request):
     if request.user.is_authenticated:
-        return render(request, 'rod/home.html', {'title': 'Strona główna'})
+        return render(request,'rod/home.html',{'title':'Strona główna'})
     else:
         return redirect('/')
 
@@ -66,7 +65,6 @@ def drones(request):
     else:
         return redirect('/')
 
-
 def new_drone(request):
     if request.user.is_authenticated:
         if request.method == 'POST':
@@ -81,60 +79,93 @@ def new_drone(request):
         return render(request, 'rod/new_drone.html', {'title': 'Drony_nowe'})
 
 
+
 def updateDron(request, pk):
-    dron = Drony.objects.get(id_drona=pk)
-    form = DronUpdateForm(instance=dron)
+    if request.user.is_authenticated:
+        dron = Drony.objects.get(id_drona = pk)
+        form = DronUpdateForm(instance = dron)
 
-    if request.method == 'POST':
-        form = DronUpdateForm(request.POST, instance=dron)
-        if form.is_valid():
-            form.save()
-            return redirect('/drones')
-    context = {'form': form}
-    return render(request, 'rod/dron_update.html', context)
-
-
-def delDron(request, pk):
-    dron = Drony.objects.get(id_drona=pk)
-    if request.method == "POST":
-        dron.delete()
+        if request.method == 'POST':
+            form = DronUpdateForm(request.POST, instance=dron)
+            if form.is_valid():
+                form.save()
+                return redirect('/drones')
+        context = {'form':form}
+        return render(request, 'rod/dron_update.html', context)
+    else:
         return redirect('/')
-    context = {'dron': dron}
-    return render(request, 'rod/dron_confirm_delete.html', context)
 
+def delDron(request,pk):
+    if request.user.is_authenticated:
+        dron = Drony.objects.get(id_drona = pk)
+        if request.method == "POST":
+                dron.delete()
+                return redirect('/')
+        context = {'dron':dron}
+        return render(request, 'rod/dron_confirm_delete.html', context)
+    else:
+        return redirect('/')
 
-# -----------------------------------------------------------------------------DRONY-------------------------------------------------------------
+##-----------------------------------------------------------------------------DRONY-------------------------------------------------------------
 def map(request):
     if request.user.is_authenticated:
         zgloszenia = Zgloszenia.objects.all()
-        return render(request, 'rod/map.html', {'title': 'Mapa', 'zgloszenia': zgloszenia})
+        return render(request,'rod/map.html',{'title':'Mapa', 'zgloszenia':zgloszenia})
     else:
         return redirect('/')
-
 
 def shedule(request):
     if request.user.is_authenticated:
-        return render(request, 'rod/shedule.html', {'title': 'Rozkład jazdy'})
+        zgloszenia = Zgloszenia.objects.filter(zarejestrowane=False).order_by('-id_zgloszenia')
+        for row in zgloszenia:
+            row.zdjecie = b64encode(row.zdjecie).decode('ascii')
+
+        paginator = Paginator(zgloszenia, 1)
+        page_number = request.GET.get('page')
+        page_obj = paginator.get_page(page_number)
+        return render(request, 'rod/last_reports.html', {'title' : 'Ostatnie zgłoszenia', 'zgloszenia':page_obj})
     else:
         return redirect('/')
-
 
 def all_archive(request):
     if request.user.is_authenticated:
-        return render(request, 'rod/all_archive.html', {'tittle': 'Wszystkie zgłoszenia'})
+        zgloszenia = Zgloszenia.objects.get_queryset().order_by('-id_zgloszenia')
+        for row in zgloszenia:
+            row.zdjecie = b64encode(row.zdjecie).decode('ascii')
+        paginator = Paginator(zgloszenia, 1)
+        page_number = request.GET.get('page')
+        page_obj = paginator.get_page(page_number)
+
+        return render(request,'rod/all_archive.html',{'tittle':'Wszystkie zgłoszenia', 'zgloszenia':page_obj})
     else:
         return redirect('/')
-
 
 def interv_archive(request):
     if request.user.is_authenticated:
-        return render(request, 'rod/interv_archive.html', {'title': 'Zgłoszenia wymagające interwencji'})
+        zgloszenia = Zgloszenia.objects.filter(zgloszenie_sluzbom=True, zarejestrowane=True).order_by('-id_zgloszenia')
+        for row in zgloszenia:
+            row.zdjecie = b64encode(row.zdjecie).decode('ascii')
+        paginator = Paginator(zgloszenia, 1)
+        page_number = request.GET.get('page')
+        page_obj = paginator.get_page(page_number)
+
+        return render(request, 'rod/interv_archive.html',{'title':'Zgłoszenia wymagające interwencji', 'zgloszenia':page_obj})
     else:
         return redirect('/')
-
 
 def noninterv_archive(request):
     if request.user.is_authenticated:
-        return render(request, 'rod/noninterv_archive.html', {'title': 'Zgłoszenia niewymagające interwencji'})
+        zgloszenia = Zgloszenia.objects.filter(zgloszenie_sluzbom=False, zarejestrowane=True).order_by('-id_zgloszenia')
+        for row in zgloszenia:
+            row.zdjecie = b64encode(row.zdjecie).decode('ascii')
+        paginator = Paginator(zgloszenia, 1)
+        page_number = request.GET.get('page')
+        page_obj = paginator.get_page(page_number)
+
+
+        return render(request,'rod/noninterv_archive.html',{'title':'Zgłoszenia niewymagające interwencji','zgloszenia':page_obj})
     else:
         return redirect('/')
+
+
+
